Restructure JSON serialization code
