+++ palette.h
@@ -5,6 +5,7 @@
 #include <cstdint>
 #include <array>
 #include <algorithm>
+#include "json.hpp"
 #include "pixeltypes.h"
 
 class Palette
@@ -93,3 +94,29 @@ public:
     }
 };
 
+inline void to_json(nlohmann::json& j, const Palette & palette) 
+{
+    auto colorsJson = nlohmann::json::array();
+    for (const auto& color : palette.getColors()) 
+        colorsJson.push_back(color);  // Uses CRGB serializer
+        
+    j = 
+    {
+        {"colors", colorsJson},
+        {"blend", palette._bBlend}
+    };
+}
+
+inline void from_json(const nlohmann::json& j, std::unique_ptr<Palette>& palette) 
+{
+    // Deserialize the "colors" array
+    std::vector<CRGB> colors;
+    for (const auto& colorJson : j.at("colors")) 
+        colors.push_back(colorJson.get<CRGB>()); // Use CRGB's from_json function
+
+    // Deserialize the "blend" flag, defaulting to true if not present
+    bool blend = j.value("blend", true);
+
+    // Create new Palette
+    palette = std::make_unique<Palette>(colors, blend);
+}
